requirements.txt
Flask
requests
feedparser
Pillow
bing-image-downloader
edge-tts 
gunicorn
groq